+++ src/app/move-popup/sockets.component.js
@@ -1,3 +1,4 @@
+import _ from 'underscore';
 import template from './sockets.html';
 import './sockets.scss';
 
@@ -16,4 +17,18 @@ function SocketsCtrl($i18next) {
   const vm = this;
 
   vm.bestRatedText = `\n${$i18next.t('DtrReview.BestRatedTip')}`;
+
+  vm.anyBestRatedUnselected = function(category) {
+    let anyBestRatedAndUnselected = false;
+
+    _.each(category.sockets, (socket) => {
+      _.each(socket.plugOptions, (plugOption) => {
+        if ((plugOption !== socket.plug) && plugOption.bestRated) {
+          anyBestRatedAndUnselected = true;
+        }
+      });
+    });
+
+    return anyBestRatedAndUnselected;
+  };
 }